Renamed launcher module to algorithms.
