+++ phy/cluster/__init__.py
@@ -3,7 +3,7 @@
 
 """Automatic and manual clustering facilities."""
 
-from .launcher import run
+from .algorithms import run
 from .session import Session
 from .view_models import (BaseClusterViewModel,
                           HTMLClusterViewModel,
