+++ phy/cluster/session.py
@@ -17,7 +17,7 @@ from ..io.base import BaseSession
 from ..io.kwik.model import KwikModel
 from ..io.kwik.store_items import create_store
 from .manual.gui import ClusterManualGUI
-from .launcher import KlustaKwik
+from .algorithms import KlustaKwik
 
 
 #------------------------------------------------------------------------------
